+++ src/entities/Player.java
@@ -4,7 +4,6 @@ import java.awt.Graphics;
 import java.awt.image.BufferedImage;
 
 import main.Game;
-import utilz.HelpMethods;
 import utilz.LoadSave;
 
 
@@ -19,8 +18,13 @@ public class Player extends Entity {
 
 	private int playerAction = IDLE;
 	private boolean moving = false, attacking = false;
-    private boolean up, down, left, right;
-    private float playerSpeed = 1.0f;
+    private boolean up, down, left, right, jump;
+    private float playerSpeed = 0.5f * Game.SCALE;
+    private float airSpeed = 0f;
+    private float gravity = 0.045f * Game.SCALE;
+    private float jumpSpeed = -2.55f * Game.SCALE;
+    private float fallSpeedAfterCollision = 0.1f * Game.SCALE;
+    private boolean inAir = false;
 
     private int[][] levelData;
 
@@ -30,7 +34,7 @@ public class Player extends Entity {
         super(x, y, width, height);
         loadAnimations();
 
-        initHitbox(x, y, 18 * Game.SCALE, 32 * Game.SCALE);
+        initHitbox(x, y, (int)(18 * Game.SCALE), (int)(32 * Game.SCALE));
     }
 
     public void update() {
@@ -66,6 +70,13 @@ public class Player extends Entity {
 		} else {
 			playerAction = IDLE;
 		}
+        if (inAir){
+            if (airSpeed < 0) {
+                playerAction = JUMPING;
+            } else {
+                playerAction = FALLING;
+            }
+        }
 
         if (attacking) {
             playerAction = ATTACKING;
@@ -83,34 +94,71 @@ public class Player extends Entity {
 
     private void updatePosition() {
         moving = false;
-		if (!left && !right && !up && !down)
+
+        if (jump) {
+            jump();
+        }
+		if (!left && !right && !inAir) {
 			return;
+        }
 
-		float xSpeed = 0, ySpeed = 0;
+		float xSpeed = 0;
 
-		if (left && !right)
-			xSpeed = -playerSpeed;
-		else if (right && !left)
-			xSpeed = playerSpeed;
+		if (left) {
+			xSpeed  -= playerSpeed;
+        }
+		if (right) {
+            xSpeed += playerSpeed;
+        }
 
-		if (up && !down)
-			ySpeed = -playerSpeed;
-		else if (down && !up)
-			ySpeed = playerSpeed;
+        if (!inAir) {
+            if (!isEntityOnFloor(hitbox, levelData)) {
+                inAir = true;
+            }
+        }
 
-		// if (canMoveHere(x + xSpeed, y + ySpeed, width, height, levelData)) {
-		// 	this.x += xSpeed;
-		// 	this.y += ySpeed;
-		// 	moving = true;
-		// }
+        if (inAir) {
+            if (canMoveHere(hitbox.x, hitbox.y + airSpeed, hitbox.width, hitbox.height, levelData)) {
+                hitbox.y += airSpeed;
+                airSpeed += gravity;
+                updateXPosition(xSpeed);
+            } else {
+                hitbox.y = getEntityYPositionUnderRoofOrAboveFloor(hitbox, airSpeed);
+                if (airSpeed > 0){
+                    resetInAir();
+                }
+                else{
+                   airSpeed = fallSpeedAfterCollision;
+                }
+                updateXPosition(xSpeed);;
+            }
 
-        if (canMoveHere(hitbox.x + xSpeed, hitbox.y + ySpeed, hitbox.width, hitbox.height, levelData)) {
-			hitbox.x += xSpeed;
-			hitbox.y += ySpeed;
-			moving = true;
-		}
-
+        }else {
+            updateXPosition(xSpeed);
+        }
+        moving = true;
 	}
+
+    private void jump() {
+        if (inAir) {
+            return;
+        }
+        inAir = true;
+        airSpeed = jumpSpeed;
+    }
+
+    private void resetInAir() {
+        inAir = false;
+        airSpeed = 0;
+    }
+
+    private void updateXPosition(float xSpeed) {
+        if (canMoveHere(hitbox.x + xSpeed, hitbox.y, hitbox.width, hitbox.height, levelData)) {
+			hitbox.x += xSpeed;
+		} else {
+            hitbox.x = getEntityXPositionNextToWall(hitbox, xSpeed, levelData);
+        }
+    }
 
     private void loadAnimations() {
         BufferedImage image = LoadSave.getSpriteAtlas(PLAYER_ATLAS);
@@ -125,6 +173,9 @@ public class Player extends Entity {
     
     public void loadLevelData(int[][] levelData) {
         this.levelData = levelData;
+        if (!isEntityOnFloor(hitbox, levelData)){
+            inAir = true;
+        }
     }
     public void resetDirectionBooleans() {
         up = false;
@@ -167,5 +218,9 @@ public class Player extends Entity {
 
     public void setRight(boolean right) {
         this.right = right;
+    }
+
+    public void setJump(boolean jump) {
+        this.jump = jump;
     }
 }