+++ src/gamestates/State.java
@@ -1,0 +1,23 @@
+package gamestates;
+
+import java.awt.event.MouseEvent;
+
+import main.Game;
+import ui.MenuButton;
+
+public class State {
+
+    protected Game game;
+
+    public State(Game game) {
+        this.game = game;
+    }
+
+    public boolean isIn(MouseEvent e, MenuButton menuButton) {
+        return menuButton.getBounds().contains(e.getPoint());
+    }
+
+    public Game getGame() {
+        return game;
+    }
+}